# Speed up all the array.fs for loops