--- a/src/fsharp/FSharp.Core/array.fs
+++ b/src/fsharp/FSharp.Core/array.fs
@@ -556,7 +556,6 @@ namespace Microsoft.FSharp.Collections
             | 0 -> empty
             | resLen -> Array.subUnchecked i resLen array
 
-
         [<CompiledName("FindBack")>]
         let findBack f (array: _[]) =
             checkNonNull "array" array