This addresses issue #1419 , where a pattern was used in for loops that leads to bad performance. This PR should speed up many functions in array.fs.

`array.rev` was also slightly refactored at the same time to make it much faster.