@jackmott 2016-08-11
ci error looks related to AsyncChoice, which I don't think I touched.
Related to issue #1379 ?