@jackmott 2016-08-11
I'm merging the latest master into this and cleaning some things up. Please don't merge yet.