@dsyme 2016-08-12
This looks really good.  I've left some comments about places where ``let len = ...`` can be removed.  

We should accept this into F# regardless of whether the JIT gets fixed to address https://github.com/Microsoft/visualfsharp/issues/1419 (because FSHarp.Core.dll often gets used with current versions of the JIT)