@jackmott 2016-08-12
Thanks for the review @dsyme, I'll check those cases and make adjustments.