@jackmott 2016-08-12
That appveyor error looks to be unrealted, I think.