@jackmott 2016-08-12
Yeah it looks like it is that AsyncChoice thing that comes up sometimes.