@liboz 2016-08-14
Are there any numbers for how much of a difference this change makes?