@jackmott 2016-08-15
Sure. Comparing a pretty average case of a single function  `map` with a simple mapping function `(fun x -> x + 1)` the perf difference is below.  As the mapping function gets more instructions, the % difference will tend towards zero of course.

A small difference but the code gets simpler and it affects many of the fundamental functions like subUnchecked, map, create, and init. The core DLL also gets 1,024 bytes smaller.


```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240908 ticks, Resolution=446.2477 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=SIMDBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```
 Method |  Length |            Median |         StdDev | Scaled | Gen 0 | Gen 1 |  Gen 2 | Bytes Allocated/Op |
------- |-------- |------------------ |--------------- |------- |------ |------ |------- |------------------- |
    **Old** |      **10** |        **17.5030 ns** |      **0.5275 ns** |   **1.00** |  **0.05** |     **-** |      **-** |              **26.88** |
    New |      10 |        14.1205 ns |      0.4858 ns |   0.81 |  0.05 |     - |      - |              26.11 |
    **Old** |   **10000** |    **10,212.8762 ns** |    **118.7990 ns** |   **1.00** | **27.45** |     **-** |      **-** |          **15,484.71** |
    New |   10000 |     8,963.2690 ns |    329.8907 ns |   0.88 | 29.07 |     - |      - |          16,395.45 |
    **Old** | **1000000** | **2,078,782.0718 ns** | **25,255.4172 ns** |   **1.00** |     **-** |     **-** | **858.90** |       **1,854,656.87** |
    New | 1000000 | 1,956,501.4913 ns | 15,930.2483 ns |   0.94 |     - |     - | 832.00 |       1,806,253.44 |